improved architecture
