+++ client/src/components/LoadingScreen/recordPlayer.css
@@ -1,16 +1,4 @@
-.loading-screen-wrapper {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  background-color: var(--background-primary-color);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 1000;
-}
-
+/* From Uiverse.io by TheAbieza */ 
 .container {
   width: 175px;
   height: 175px;
