+++ client/src/components/LoadingScreen/recordPlayer.html
@@ -1,0 +1,17 @@
+<!-- From Uiverse.io by TheAbieza --> 
+<div class="container">
+  <div class="plate">
+      <div class="black">
+          <div class="border">
+              <div class="white">
+                  <div class="center"></div>
+              </div>
+          </div>
+      </div>
+  </div>
+
+  <div class="player">
+      <div class="rect"></div>
+      <div class="circ"></div>
+  </div>
+</div>
